+++ test/intl/Intl__NumberFormatTest.mjs
@@ -15,7 +15,7 @@ var currencyFormatter = new Intl.NumberFormat(["fr-FR"], {
 console.log(Intl.NumberFormat.supportedLocalesOf([
           "fr-FR",
           "en-US"
-        ], undefined));
+        ]));
 
 console.log(currencyFormatter.format(123.23));